+++ src/SampleApi/Modules/Users/Domain/User.Events.cs
@@ -4,7 +4,7 @@ namespace VolcanoBlue.SampleApi.Modules.Users.Domain
 {
     public sealed partial class User
     {
-        public sealed record UserCreated(int EventId, Guid Id, string Name, string Email) : IEvent;
-        public sealed record UserEmailChanged(int EventId, Guid Id, string NewEmail) : IEvent;
+        public sealed record UserCreated(long EventId, Guid Id, string Name, string Email) : IEvent;
+        public sealed record UserEmailChanged(long EventId, Guid Id, string NewEmail) : IEvent;
     }
 }
